fix: NEWSページのロジックをhookに切り出し
